requirements.txt
arabic-reshaper==2.1.4
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.1.1
Django==4.1.4
django-appconf==1.0.5
django-compressor==4.0
django-filter==21.1
django-htmlmin==0.11.0
django-storages==1.13.2
djangorestframework==3.12.4
et-xmlfile==1.1.0
future==0.18.2
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
html5lib==1.1
idna==3.4
jdcal==1.4.1
Markdown==3.3.4
mysqlclient==1.4.6
openpyxl==3.0.9
Pillow==9.1.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2022.1
rcssmin==1.1.0
reportlab==3.6.1
requests==2.28.1
rjsmin==1.2.0
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.7
urllib3==1.26.13
webencodings==0.5.1
wincertstore==0.2
xhtml2pdf==0.2.5
